fix: 상품 상세 페이지 파일 수정
