+++ src/pages/ProductDetail/index.tsx
@@ -1,12 +1,13 @@
-import { Box, Button, Container, Image, Select, Spinner, Text } from '@chakra-ui/react';
+import { Box, Container, Spinner, Text } from '@chakra-ui/react';
 import { useState } from 'react';
 import { Navigate, useParams } from 'react-router-dom';
 
 import { useGetProductDetail } from '@/api/hooks/useGetProductDetail';
 import { useGetProductOptions } from '@/api/hooks/useGetProductOptions';
-import { useAuth } from '@/provider/Auth';
-import { getDynamicPath, RouterPath } from '@/routes/path';
-import type { ProductOption } from '@/types';
+import { GiftButton } from '@/components/features/ProductDetail/GiftButton';
+import { ProductInfo } from '@/components/features/ProductDetail/ProductInfo';
+import { ProductOptions } from '@/components/features/ProductDetail/ProductOptions';
+import { RouterPath } from '@/routes/path';
 
 export const ProductDetailPage = () => {
   const { productId = '' } = useParams<{ productId: string }>();
@@ -20,7 +21,6 @@ export const ProductDetailPage = () => {
     isLoading: isLoadingOptions,
     isError: isErrorOptions,
   } = useGetProductOptions(productId);
-  const authInfo = useAuth();
   const [quantity, setQuantity] = useState(1);
 
   if (isLoadingDetail || isLoadingOptions) return <Spinner />;
@@ -28,29 +28,20 @@ export const ProductDetailPage = () => {
     return <Navigate to={RouterPath.home} />;
 
   const handlePurchase = () => {
-    if (!authInfo) {
-      return window.location.replace(getDynamicPath.login());
-    }
-
     // Handle purchase logic here
   };
 
   return (
     <Container>
       <Box>
-        <Image src={productDetail.imageURL} alt={productDetail.name} />
-        <Text>{productDetail.brandInfo.name}</Text>
-        <Text>{productDetail.name}</Text>
-        <Text>{productDetail.price.sellingPrice.toLocaleString()}원</Text>
+        <ProductInfo productDetail={productDetail} />
         <Text>카톡 친구가 아니어도 선물 코드로 선물 할 수 있어요!</Text>
-        <Select value={quantity} onChange={(e) => setQuantity(parseInt(e.target.value, 10))}>
-          {productOptions.map((option: ProductOption, index: number) => (
-            <option key={index} value={option.quantity}>
-              {option.quantity}
-            </option>
-          ))}
-        </Select>
-        <Button onClick={handlePurchase}>나에게 선물하기</Button>
+        <ProductOptions
+          productOptions={productOptions}
+          quantity={quantity}
+          setQuantity={setQuantity}
+        />
+        <GiftButton handlePurchase={handlePurchase} />
       </Box>
     </Container>
   );
